Added check if file with printers exists.

+++ printers/report.sh
@@ -1,8 +1,13 @@
 #!/bin/bash
 
+printersfile="printers.txt"
 today=`date +%Y-%m-%d`
 reportfile=$today"_printers_report.csv"
 
+if ! [ -f $printersfile ]; then
+	echo "File $printersfile not exists."
+	exit 1
+fi
 
 
 echo "Printer; counter" | tee > $reportfile
